CP-54351 Move pluginutil from /opt
